refactor：storybook：component：PersonForm：関数切り出し
