+++ src/stories/components/organisms/Forms/PersonForms/index.tsx
@@ -5,16 +5,19 @@ import {
   LogoutDefaultFlex,
   LogoutDefaultFlexType,
 } from '@/stories/components/molecules/Button/LogoutDefaultFlex';
-import { Form, FormType } from '@/stories/components/molecules/Form/PersonForm';
+import {
+  PersonForm,
+  PersonFormType,
+} from '@/stories/components/molecules/Form/PersonForm';
 
-export type PersonFormsType = FormType & LogoutDefaultFlexType;
+export type PersonFormsType = PersonFormType & LogoutDefaultFlexType;
 export const PersonForms: FC<PersonFormsType> = ({
   processResultFunc,
   authHandler,
 }: PersonFormsType) => {
   return (
     <Box>
-      <Form processResultFunc={processResultFunc} />
+      <PersonForm processResultFunc={processResultFunc} />
       <LogoutDefaultFlex authHandler={authHandler} />
     </Box>
   );
